requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.0.8
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.5.0
djoser==2.0.5
gunicorn==20.0.4
Markdown==3.2.2
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
